+++ crates/pallets/starknet/src/lib.rs
@@ -66,11 +66,11 @@ use blockifier::execution::entry_point::{
     CallEntryPoint, CallInfo, CallType, EntryPointExecutionContext, ExecutionResources,
 };
 use blockifier::execution::errors::{EntryPointExecutionError, PreExecutionError};
-use blockifier::state::cached_state::ContractStorageKey;
 use blockifier_state_adapter::BlockifierStateAdapter;
 use frame_support::pallet_prelude::*;
 use frame_support::traits::Time;
 use frame_system::pallet_prelude::*;
+use itertools::Itertools;
 use mp_block::state_update::StateUpdateWrapper;
 use mp_block::{Block as StarknetBlock, Header as StarknetHeader};
 use mp_contract::ContractAbi;
@@ -120,6 +120,7 @@ macro_rules! log {
 
 #[frame_support::pallet]
 pub mod pallet {
+    use mp_block::state_update::StorageDiffWrapper;
     use mp_contract::class::{ClassUpdateWrapper, ContractClassData, ContractClassWrapper};
 
     use super::*;
@@ -216,13 +217,22 @@ pub mod pallet {
     fn store_state_update<T: Config>(encoded_data: &Vec<u8>) {
         match StateUpdateWrapper::decode(&mut encoded_data.as_slice()) {
             Ok(state_update) => {
-                for (address, storage_diffs) in state_update.state_diff.storage_diffs {
-                    for storage_diff in storage_diffs {
-                        let contract_storage_key: ContractStorageKey =
-                            (ContractAddress(address.into()), StorageKey(storage_diff.key.into()));
-                        let value = StarkFelt(storage_diff.value.into());
-                        <StorageView<T>>::insert(contract_storage_key, value)
-                    }
+                for (contract_address, storage_diffs) in state_update.state_diff.storage_diffs {
+                    let contract_address = ContractAddress(contract_address.into());
+
+                    // Previous storage has to be persisted. Since we have no way of using hashmaps
+                    // in a no-std environment, we are currently joining previous and current
+                    // storage updates and filtering out duplicates
+                    // TODO: find a more performant way to do this.
+                    let storage_old = <StorageView<T>>::get(contract_address);
+                    let storage_new: Vec<(StorageKey, StarkFelt)> = storage_diffs
+                        .into_iter()
+                        .map(|StorageDiffWrapper { key, value }| (StorageKey::from(key), StarkFelt::from(value)))
+                        .collect();
+                    let storage: Vec<(StorageKey, StarkFelt)> =
+                        storage_new.into_iter().chain(storage_old.into_iter()).unique_by(|(key, _)| *key).collect();
+
+                    <StorageView<T>>::insert(contract_address, storage);
                 }
 
                 // nonces stored for accessing on `starknet_getNonce` RPC call.
@@ -371,12 +381,13 @@ pub mod pallet {
     #[pallet::getter(fn nonce)]
     pub(super) type Nonces<T: Config> = StorageMap<_, Identity, ContractAddress, Nonce, ValueQuery>;
 
-    /// Mapping from Starknet contract storage key to its value.
+    /// Mapping from Starknet contract storage to its keys and values.
     /// Safe to use `Identity` as the key is already a hash.
     #[pallet::storage]
     #[pallet::unbounded]
     #[pallet::getter(fn storage)]
-    pub(super) type StorageView<T: Config> = StorageMap<_, Identity, ContractStorageKey, StarkFelt, ValueQuery>;
+    pub(super) type StorageView<T: Config> =
+        StorageMap<_, Identity, ContractAddress, Vec<(StorageKey, StarkFelt)>, ValueQuery>;
 
     /// The last processed Ethereum block number for L1 messages consumption.
     /// This is used to avoid re-processing the same Ethereum block multiple times.
@@ -423,7 +434,7 @@ pub mod pallet {
         /// a test environment or in the case of a migration of an existing chain state.
         pub contracts: Vec<(ContractAddress, SierraClassHash)>,
         pub sierra_to_casm_class_hash: Vec<(SierraClassHash, CasmClassHash)>,
-        pub storage: Vec<(ContractStorageKey, StarkFelt)>,
+        pub storage: Vec<(ContractAddress, Vec<(StorageKey, StarkFelt)>)>,
         /// The address of the fee token.
         /// Must be set to the address of the fee token ERC20 contract.
         pub fee_token_address: ContractAddress,
@@ -461,7 +472,7 @@ pub mod pallet {
 
             self.storage
                 .iter()
-                .for_each(|(contract_storage_key, value)| StorageView::<T>::insert(contract_storage_key, value));
+                .for_each(|(contract_address, storage)| StorageView::<T>::insert(contract_address, storage));
 
             LastKnownEthBlock::<T>::set(None);
             // Set the fee token address from the genesis config.
@@ -1005,7 +1016,17 @@ impl<T: Config> Pallet<T> {
     pub fn get_storage_at(contract_address: ContractAddress, key: StorageKey) -> Result<StarkFelt, DispatchError> {
         // Get state
         ensure!(ContractClassHashes::<T>::contains_key(contract_address), Error::<T>::ContractNotFound);
-        Ok(Self::storage((contract_address, key)))
+
+        match Self::storage(contract_address).iter().find(|(storage_key, _)| key == *storage_key) {
+            Some((_, value)) => Ok(*value),
+            None => Err(DispatchError::CannotLookup),
+        }
+    }
+
+    /// Returns a storage keys and values of a given contract
+    pub fn get_storage_from(contract_address: ContractAddress) -> Result<Vec<(StorageKey, StarkFelt)>, DispatchError> {
+        let changes = Self::storage(contract_address);
+        Ok(changes)
     }
 
     /// Store a Starknet block in the blockchain.