+++ CHANGELOG.md
@@ -3,6 +3,8 @@ git # Deoxys Changelog
 ## Next release
 
 - fix l1 thread to reflect correct state_root, block_number, block_hash
+- fix: remove gas_price and update starknet-rs from fork (temporary fix)
+- fix(root): got state root to work (does not support class root yet)
 - refactor(substrate_hash): Substrate hash is now retrieved via rpc client in
   `l2.rs`
 - fix(worflows): fix toolchain and cache issue